refactor: Update webserver.js to fix date calculation bug
This commit fixes a bug in the webserver.js file where the calculation of the week ago date was incorrect. The bug was causing incorrect data retrieval for the "/messages/week" endpoint. The fix ensures that the correct date is used for fetching messages from the last week, improving the accuracy of the data returned by the endpoint.